WIP: Replace M68k core w/ Musashi 680x0
The entire point of this was that I eventually wanted to add a GDB stub
in the emulator so I can debug the VIA, and I didn't want to waste that
effort w/ a CPU core I wasn't planning on using. So, now, it's changed.
The system right now gets up to drawing the checkerboard background and,
supposedly, attempting to draw the blinking disk icon. Interrupts are
probably broken. Hard to tell.

Also, the new CPU core gives us the M68030 and M68040 and the FPU all
for *free*. So yeah, expect some SE/30 action coming soon. Ish.